+++ CHANGELOG.md
@@ -32,7 +32,7 @@ NOTICE: Old saves will not work with this update!
 - Added a short sound effect when completing a quest objective.
 - Added new rogue passive skill "Resourceful" at level 15.
   - Moved Low Blow to L18, Shield Break to L22, and removed Focus Up.
-- Added 27 items:
+- Added 29 items:
   - Rookie Spear
   - Squire Gauntlets
   - Flora's Rose
@@ -49,6 +49,7 @@ NOTICE: Old saves will not work with this update!
   - Enchanted Snake Skin
   - Rattlesnake Fang
   - Cast-Iron Pot
+  - Recall Potion
   - Dalia's Delight
   - Bronze Gauntlets
   - Smirnoff's Lightning Lance